+++ blocks/json-output-block/output-types/json.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line react/prop-types
 const Json = ({ children }) => {
   return Array.isArray(children) ? children[0] : null
 }