+++ utils/xml-output/src/index.js
@@ -1,8 +1,9 @@
 import { create } from 'xmlbuilder2'
 
+// eslint-disable-next-line react/prop-types
 export const XmlOutput = ({ children }) => {
   return children
-    ? create(children[0], { invalidCharReplacement: '' })
+    ? create({ invalidCharReplacement: '' }, children[0])
         .dec({ encoding: 'UTF-8' })
         .end()
     : null